+++ packages/extension-koni-ui/src/Popup/XcmTransfer/XcmTransfer.tsx
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { ChainRegistry, DropdownTransformOptionType, NetworkJson } from '@subwallet/extension-base/background/KoniTypes';
-import { SupportedCrossChainsMap } from '@subwallet/extension-koni-base/api/supportedCrossChains';
+import { SupportedCrossChainsMap } from '@subwallet/extension-koni-base/api/xcm/supportedCrossChains';
 import { AccountContext, ActionContext, Button, Warning } from '@subwallet/extension-koni-ui/components';
 import InputBalance from '@subwallet/extension-koni-ui/components/InputBalance';
 import LoadingContainer from '@subwallet/extension-koni-ui/components/LoadingContainer';
